+++ changelog.md
@@ -1,5 +1,7 @@
 # Unreleased
 
+* Update the files menu base dir when tabbing or inserting the current buffer's dir.
+* Add mapping `<c-b>` for the files menu, cycling the base dir for file creation.
 * Add mapping `<c-d>` for the files menu, inserting the current buffer's directory.
 
 # 0.9.9.9